refactor(nat): Move stateless NAT code to a dedicated file
Make room for stateful NAT: Move all the code dedicated to stateless NAT
into a dedicated file, in preparation for stateful NAT support.
